# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для анализа композиционного центра произведений искусства",
        "title": "Art Analysis API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/analysis_orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список заявок. Пользователь видит только свои заявки, модератор — все заявки. Неавторизованный пользователь получает ошибку доступа.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Получить список заявок на анализ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата (фильтр от)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (фильтр до)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.DTO_Resp_Order"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/analysis_orders/{id}": {
            "get": {
                "description": "Возвращает детальную информацию о заявке на анализ по её идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Получить заявку по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали заявки",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_Order"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет статус заявки и/или модератора, если это требуется",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Обновить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_OrderUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная заявка",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_Order"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Полностью удаляет заказ анализа произведения искусства",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Удалить заказ на анализ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заказа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID удалённого заказа",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_SimpleID"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/analysis_orders/{id}/form": {
            "put": {
                "description": "Переводит заявку из статуса черновика в статус сформированной",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Сформировать заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сформированная заявка",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_Order"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID or cannot form order",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/analysis_orders/{id}/resolve": {
            "put": {
                "description": "Выполняет завершение или отклонение заявки по анализу произведения искусства. При завершении рассчитывается финальный композиционный центр.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Завершить/отклонить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие с заявкой (action: complete | reject)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённая заявка",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_Order"
                        }
                    },
                    "400": {
                        "description": "Invalid input or missing required fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/experts": {
            "get": {
                "description": "Возвращает список всех экспертов. Поддерживает фильтрацию по названию.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Получить список экспертов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию эксперта (подстрока)",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.DTO_Resp_Expert"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт нового эксперта с указанными атрибутами.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Создать нового эксперта",
                "parameters": [
                    {
                        "description": "Данные нового эксперта",
                        "name": "expert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_ExpertCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ArtExpert"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании эксперта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.DTO_Req_OrderUpdate": {
            "type": "object",
            "properties": {
                "moderator_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.DTO_Resp_Expert": {
            "description": "Полная информация об эксперте",
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string",
                    "example": "geometric_center"
                },
                "description": {
                    "type": "string",
                    "example": "Определяет композиционный центр"
                },
                "id_artcenter": {
                    "type": "integer",
                    "example": 1
                },
                "img_url": {
                    "type": "string",
                    "example": "expert_photo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Иванов И.И."
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Центр анализа №1"
                }
            }
        },
        "handler.DTO_Resp_Order": {
            "description": "Полная информация о заявке с экспертами",
            "type": "object",
            "properties": {
                "date_completed": {
                    "type": "string",
                    "example": "2025-11-07T12:00:00Z"
                },
                "date_created": {
                    "type": "string",
                    "example": "2025-11-05T12:00:00Z"
                },
                "date_formed": {
                    "type": "string",
                    "example": "2025-11-06T12:00:00Z"
                },
                "experts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.DTO_Resp_OrderExpert"
                    }
                },
                "id_creator": {
                    "type": "integer",
                    "example": 1
                },
                "id_moderator": {
                    "type": "integer",
                    "example": 2
                },
                "id_order": {
                    "type": "integer",
                    "example": 5
                },
                "order_status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "formed",
                        "completed",
                        "rejected"
                    ],
                    "example": "formed"
                },
                "result_x": {
                    "type": "number",
                    "example": 0.51
                },
                "result_y": {
                    "type": "number",
                    "example": 0.48
                }
            }
        },
...
        "handler.DTO_Resp_SimpleID": {
            "description": "Ответ, содержащий только идентификатор",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}